+++ examples/basics/mnist-tfslim.py
@@ -46,7 +46,7 @@ class Model(ModelDesc):
         cost = tf.nn.sparse_softmax_cross_entropy_with_logits(logits=logits, labels=label)
         cost = tf.reduce_mean(cost, name='cross_entropy_loss')
 
-        acc = tf.to_float(tf.nn.in_top_k(logits, label, 1))
+        acc = tf.cast(tf.nn.in_top_k(logits, label, 1), tf.float32)
 
         acc = tf.reduce_mean(acc, name='accuracy')
         summary.add_moving_summary(acc)
